feat: update logout mysky logic and add logout button
