+++ src/components/Sites/routes/ResolverSkylink/components/GenerateResolverSkylink/GenerateResolverSkylink.scss
@@ -53,6 +53,20 @@
         width: 100%;
         align-items: center;
 
+        .small-text {
+          font-size: 16px;
+        }
+
+        .margin-left {
+          margin-left: 1rem;
+        }
+
+        .user-container {
+          display: flex;
+          align-items: center;
+          width: 100%;
+        }
+
         .text-input {
           flex: 1;
           padding: 8px 12px;
@@ -141,6 +155,20 @@
             8px,
             var(--themeColorDark),
             var(--primaryDisabled)
+          );
+          margin-top: 1rem;
+          margin-bottom: 1rem;
+        }
+
+        .logout-mysky-button {
+          @include primaryButton(
+            white,
+            var(--red),
+            0.2rem 1rem,
+            none,
+            8px,
+            var(--redDark),
+            var(--redDisabled)
           );
           margin-top: 1rem;
           margin-bottom: 1rem;
